requirements.txt
pandas
requests
beautifulsoup4
playwright
python-dotenv
unidecode
fuzzywuzzy